- General Onyx optimizations, use of C instead of ObjC. - O2Encoder_TIFF optimizations, alpha support - O2Decoder_TIFF optimizations - O2Context_builtin clip rect fixed to do intersection of consecutive clips - O2Pattern work - Removed kO2ColorSpaceModelPlatformRGB and gamma adjustment

+++ CoreGraphics/CGAffineTransform.h
@@ -20,17 +20,17 @@ typedef struct {
 
 COREGRAPHICS_EXPORT const CGAffineTransform CGAffineTransformIdentity;
 
-COREGRAPHICS_EXPORT CGAffineTransform CGAffineTransformMake(float a,float b,float c,float d,float tx,float ty);
-COREGRAPHICS_EXPORT CGAffineTransform CGAffineTransformMakeRotation(float radians);
-COREGRAPHICS_EXPORT CGAffineTransform CGAffineTransformMakeScale(float scalex,float scaley);
-COREGRAPHICS_EXPORT CGAffineTransform CGAffineTransformMakeTranslation(float tx,float ty);
+COREGRAPHICS_EXPORT CGAffineTransform CGAffineTransformMake(CGFloat a,CGFloat b,CGFloat c,CGFloat d,CGFloat tx,CGFloat ty);
+COREGRAPHICS_EXPORT CGAffineTransform CGAffineTransformMakeRotation(CGFloat radians);
+COREGRAPHICS_EXPORT CGAffineTransform CGAffineTransformMakeScale(CGFloat scalex,CGFloat scaley);
+COREGRAPHICS_EXPORT CGAffineTransform CGAffineTransformMakeTranslation(CGFloat tx,CGFloat ty);
 
 COREGRAPHICS_EXPORT CGAffineTransform CGAffineTransformConcat(CGAffineTransform xform,CGAffineTransform append);
 COREGRAPHICS_EXPORT CGAffineTransform CGAffineTransformInvert(CGAffineTransform xform);
 
-COREGRAPHICS_EXPORT CGAffineTransform CGAffineTransformRotate(CGAffineTransform xform,float radians);
-COREGRAPHICS_EXPORT CGAffineTransform CGAffineTransformScale(CGAffineTransform xform,float scalex,float scaley);
-COREGRAPHICS_EXPORT CGAffineTransform CGAffineTransformTranslate(CGAffineTransform xform,float tx,float ty);
+COREGRAPHICS_EXPORT CGAffineTransform CGAffineTransformRotate(CGAffineTransform xform,CGFloat radians);
+COREGRAPHICS_EXPORT CGAffineTransform CGAffineTransformScale(CGAffineTransform xform,CGFloat scalex,CGFloat scaley);
+COREGRAPHICS_EXPORT CGAffineTransform CGAffineTransformTranslate(CGAffineTransform xform,CGFloat tx,CGFloat ty);
 
 COREGRAPHICS_EXPORT CGPoint CGPointApplyAffineTransform(CGPoint point, CGAffineTransform xform);
 COREGRAPHICS_EXPORT CGSize CGSizeApplyAffineTransform(CGSize size, CGAffineTransform xform);
